+++ rueckgrat/hub/app/main.py
@@ -14,7 +14,7 @@ from fastapi.security import HTTPBearer
 from pydantic import BaseModel
 from datetime import datetime, timezone
 from contextlib import asynccontextmanager
-from app.utils import ChatDB, Infrastructure, ImageType
+from app.utils import ChatDB, Infrastructure, ImageType, Skills
 from app.jobs import JobQueue, MetaJob, AssistantImageJob, ContactGeneratorJob
 from argon2 import PasswordHasher
 from argon2.exceptions import VerifyMismatchError, InvalidHashError
@@ -28,6 +28,7 @@ logger = Logger(__name__).get_logger()
 async def lifespan(app: FastAPI):
     random.seed(time.time())
     
+    app.state.skills = Skills("/hub/skills")
     app.state.infrastructure = Infrastructure()
     db_path = "/hub/db/chat.db"
     app.state.db = ChatDB(db_path)    
@@ -352,7 +353,7 @@ async def websocket_endpoint(websocket: WebSocket, username: str = Depends(get_c
 
                 if "chat" in data:
                     chat_request = ChatRequest(**data["chat"])
-                    job = MetaJob(user_id, chat_request, app.state.db, app.state.infrastructure)
+                    job = MetaJob(user_id, chat_request, app.state.db, app.state.infrastructure, app.state.skills)
                     app.state.job_queue.add(job)
                 elif "generate" in data:
                     generate = data["generate"]